+++ YouTubeAnalitycs/src/app/(auth)/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from "next";
 
 export const metadata: Metadata = {
-  title: "Authentication - ShadcnStore",
+  title: "Authentication - Beem Analytics",
   description: "Sign in to your account or create a new one",
 };